requirements.txt
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
ttkbootstrap>=1.10.1
pymupdf>=1.24.0
